brain tree validation

+++ ParkHereModel/src/messages/ListingReviewMessage.java
@@ -4,7 +4,7 @@ import model.Reservation;
 
 public class ListingReviewMessage extends Message{
 	public Reservation reservation;
-	public int rate;
+	public double rate;
 	public String comment;
 	public Long userId;
 }
